# Unique-Finance-Tracker/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QueueLess Pulse - Smart Queue Management for Banks
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  phoneNumber?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  staff: "staff",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phoneNumber?: string | null;
  role: UserRole;
  karma: number;
}

export type BranchPulsePulseLevel =
  (typeof BranchPulsePulseLevel)[keyof typeof BranchPulsePulseLevel];

export const BranchPulsePulseLevel = {
  calm: "calm",
  moderate: "moderate",
  busy: "busy",
  packed: "packed",
} as const;

export interface BranchPulse {
  id: number;
  name: string;
  address: string;
  city: string;
  lat: number;
  lng: number;
  queueLength: number;
  avgWaitMinutes: number;
  pulseLevel: BranchPulsePulseLevel;
  openCounters: number;
  nowServing?: string | null;
}

export interface Service {
  id: number;
  name: string;
  code: string;
  avgDurationMinutes: number;
  icon: string;
  description: string;
}

export interface ChecklistItem {
  key: string;
  label: string;
  required: boolean;
  hint?: string;
}

export interface SmartSlot {
  timeSlot: string;
  expectedWaitMinutes: number;
  score: number;
  label: string;
  recommended?: boolean;
}

export type CreateBookingInputPriority =
  (typeof CreateBookingInputPriority)[keyof typeof CreateBookingInputPriority];

export const CreateBookingInputPriority = {
  normal: "normal",
  senior: "senior",
  pregnant: "pregnant",
  disabled: "disabled",
} as const;

export interface CreateBookingInput {
  branchId: number;
  serviceId: number;
  bookingDate: string;
  timeSlot: string;
  groupSize?: number;
  priority?: CreateBookingInputPriority;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  booked: "booked",
  waiting: "waiting",
  serving: "serving",
  completed: "completed",
  no_show: "no_show",
  cancelled: "cancelled",
} as const;

export type BookingPriority =
  (typeof BookingPriority)[keyof typeof BookingPriority];

export const BookingPriority = {
  normal: "normal",
  senior: "senior",
  pregnant: "pregnant",
  disabled: "disabled",
} as const;

export interface Booking {
  id: number;
  userId: number;
  branchId: number;
  branchName: string;
  serviceId: number;
  serviceName: string;
  bookingDate: string;
  timeSlot: string;
  tokenNumber: string;
  status: BookingStatus;
  groupSize: number;
  priority: BookingPriority;
  checklistCompleted: number;
  checklistTotal: number;
  completedItems: string[];
  createdAt: string;
}

export interface QueueTracking {
  booking: Booking;
  currentlyServing: string | null;
  peopleAhead: number;
  estimatedWaitMinutes: number;
  leaveNowAdvice: string;
  walkBufferMinutes: number;
  recentlyCompleted: string[];
}

export interface SwapListing {
  id: number;
  bookingId: number;
  ownerName: string;
  branchName: string;
  serviceName: string;
  bookingDate: string;
  timeSlot: string;
  tokenNumber: string;
  note: string;
  createdAt: string;
}

export interface AdminQueueView {
  branch: BranchPulse;
  nowServing: Booking[];
  upcoming: Booking[];
  completedToday: Booking[];
}

export type BranchStatsHourlyDistributionItem = {
  hour: string;
  count: number;
};

export interface BranchStats {
  branchId: number;
  servedToday: number;
  avgServiceMinutes: number;
  peakHour: string;
  throughputPerHour: number;
  hourlyDistribution: BranchStatsHourlyDistributionItem[];
}

export type GetLoungeDisplay200 = {
  branch: BranchPulse;
  nowServing: Booking[];
  upcoming: Booking[];
};

export type GetSmartSlotsParams = {
  branchId: number;
  serviceId: number;
  date: string;
};

export type UpdateBookingChecklistBody = {
  completedItems: string[];
};

export type CreateSwapListingBody = {
  bookingId: number;
  note?: string;
};

export type AcceptSwapBody = {
  myBookingId: number;
};

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  booked: "booked",
  waiting: "waiting",
  serving: "serving",
  completed: "completed",
  no_show: "no_show",
  cancelled: "cancelled",
} as const;

export type UpdateBookingStatusBody = {
  status: UpdateBookingStatusBodyStatus;
};